backend/requirements.txt
asgiref==3.9.1
Django==5.1.1
djangorestframework==3.15.2
flake8==7.3.0
mccabe==0.7.0
pycodestyle==2.14.0
pyflakes==3.4.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
